Fixing erlotinib and vemurafenib
Apparent error was actually a wrong bond order to do with
bonds being labeled as `ar` in mol2 files, leading to
misinterpreted total bonds. Fix was to explicitly
label bonds as single or double in the mol2 file.

Note: In the maestro files produced by epik, the
correct bond order was observed.